Fix coding standard according to checkstyle.
