+++ build.gradle
@@ -1,10 +1,8 @@
 plugins {
+    id 'checkstyle'
     id 'java'
     id 'application'
     id 'com.github.johnrengelman.shadow' version '5.1.0'
-}
-plugins {
-    id 'checkstyle'
 }
 
 checkstyle {
