+++ src/main/java/duke/Launcher.java
@@ -2,9 +2,11 @@ package duke;
 
 import javafx.application.Application;
 
+/**
+ * Class to launch GUI
+ */
 public class Launcher {
     public static void main(String[] args) {
-//        Application.launch(Duke.class, args);
         Application.launch(Main.class, args);
     }
 }
