+++ src/main/java/duke/Main.java
@@ -1,8 +1,7 @@
 package duke;
-
 import java.io.IOException;
 
-import duke.JavaFXcomponent.MainWindow;
+import duke.component.MainWindow;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
