+++ src/main/java/duke/component/DialogBox.java
@@ -1,4 +1,4 @@
-package duke.JavaFXcomponent;
+package duke.component;
 
 import java.io.IOException;
 import java.util.Collections;
