+++ src/main/java/duke/component/MainWindow.java
@@ -1,7 +1,8 @@
-package duke.JavaFXcomponent;
+package duke.component;
+import static duke.parser.ProcessUserInput.process;
 
 import duke.Duke;
-import duke.Storage.DukeDecoder;
+import duke.storage.DukeDecoder;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
@@ -9,8 +10,6 @@ import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
-
-import static duke.Parser.ProcessUserInput.process;
 
 /**
  * Controller for MainWindow. Provides the layout for the other controls.
