+++ src/main/java/duke/parser/ProcessUserInput.java
@@ -1,12 +1,17 @@
-package duke.Parser;
-
-import duke.Exception.DukeException;
-import duke.Storage.DukeEncoder;
-import duke.TaskList.*;
-import Ui.Constants;
-
+package duke.parser;
 import java.util.ArrayList;
 
+import duke.exception.DukeException;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.TaskOperation;
+import duke.task.ToDo;
+import duke.ui.Constants;
+
+/**
+ * Represents methods process user input into command.
+ */
 public class ProcessUserInput {
     /**
      * Converts user command.
@@ -19,7 +24,7 @@ public class ProcessUserInput {
         int index;
         switch (typeOfTask) {
         case Constants.LIST:
-            return duke.TaskList.TaskOperation.listItems(workList);
+            return duke.task.TaskOperation.listItems(workList);
         case Constants.UNMARK:
             try {
                 userInput.substring(8);
@@ -85,8 +90,8 @@ public class ProcessUserInput {
         case Constants.DELETE:
             try {
                 userInput.substring(8);
-                index =  Integer.parseInt(userInput.split(" ")[1]);
-                return TaskOperation.delete(workList.get(index-1), workList);
+                index = Integer.parseInt(userInput.split(" ")[1]);
+                return TaskOperation.delete(workList.get(index - 1), workList);
             } catch (StringIndexOutOfBoundsException e) {
                 return new DukeException.EmptyDeleteException().throwDukeException();
 
@@ -105,7 +110,7 @@ public class ProcessUserInput {
                     System.out.println(workList.get(i).toString());
                 }
             }
-
+            // Fallthrough
         default:
             return Constants.EXIT_MESSAGE;
         }
