+++ src/main/java/duke/storage/DukeDecoder.java
@@ -1,12 +1,17 @@
-package duke.Storage;
-
-import duke.TaskList.*;
-
+package duke.storage;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.util.ArrayList;
 import java.util.Scanner;
 
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.ToDo;
+
+/**
+ * Class to initialize existing data.
+ */
 public class DukeDecoder {
     /**
      * Loads List of task from existing text file.
@@ -16,7 +21,7 @@ public class DukeDecoder {
         ArrayList<Task> workList = new ArrayList<>();
         try {
             Scanner s = new Scanner(new File("src/main/java/duke/Storage/List.txt"));
-            while (s.hasNextLine()){
+            while (s.hasNextLine()) {
                 String[] task = s.nextLine().split("\\|");
                 String typeOfTask = task[0];
                 switch (typeOfTask) {
@@ -29,6 +34,8 @@ public class DukeDecoder {
                 case "E":
                     workList.add(new Event(task[2], task[1].equals("1"), task[3]));
                     break;
+                default:
+                    // Fallthrough
                 }
             }
             s.close();
