+++ src/main/java/duke/storage/List.txt
@@ -1,1 +1,2 @@
 T|0|abc
+T|0|a
