+++ src/main/java/duke/task/TaskOperation.java
@@ -1,12 +1,13 @@
-package duke.TaskList;
-
-import Ui.Constants;
-import duke.Storage.DukeEncoder;
-
+package duke.task;
 import java.util.ArrayList;
 
-public class TaskOperation {
+import duke.storage.DukeEncoder;
+import duke.ui.Constants;
 
+/**
+ * Represents operation on task.
+ */
+public class TaskOperation {
     /**
      * Add text that user typed to the word list
      *
@@ -17,8 +18,8 @@ public class TaskOperation {
         workList.add(task);
         // Update data
         DukeEncoder.rewriteList(workList);
-        return (Constants.ARROW + "Added task: " + task.toString()) + "\n" +
-              ("Now you have " + workList.size() + " task(s) on your list.");
+        return (Constants.ARROW + "Added task: " + task.toString()) + "\n"
+                + ("Now you have " + workList.size() + " task(s) on your list.");
     }
 
     /**
@@ -31,8 +32,8 @@ public class TaskOperation {
         workList.remove(task);
         // Update data
         DukeEncoder.rewriteList(workList);
-        return (Constants.ARROW + "Deleted task: " + task.toString()) + "\n" +
-               ("Now you have " + workList.size() + " task(s) on your list.");
+        return (Constants.ARROW + "Deleted task: " + task.toString()) + "\n"
+                + ("Now you have " + workList.size() + " task(s) on your list.");
     }
 
     /**
@@ -41,7 +42,7 @@ public class TaskOperation {
     public static String listItems(ArrayList<Task> workList) {
         String toPrint = Constants.LISTING_MESSAGE + "\n";
         for (int i = 0; i < workList.size(); i++) {
-            toPrint += ((i+1) + ") " + workList.get(i).toString()) + "\n";
+            toPrint += ((i + 1) + ") " + workList.get(i).toString()) + "\n";
         }
         return toPrint;
     }
