+++ src/test/java/duke/task/TodoTest.java
@@ -1,6 +1,5 @@
-package duke.TaskList;
+package duke.task;
 
-import duke.Duke;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
